[plugin.video.science.friday] 3.0.2
