+++ plugin.video.science.friday/addon.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.science.friday" name="Science Friday" provider-name="willforde" version="3.0.1">
+<addon id="plugin.video.science.friday" name="Science Friday" provider-name="willforde" version="3.0.2">
     <requires>
         <import addon="xbmc.python" version="2.25.0"/>
-        <import addon="script.module.codequick" version="0.9.2"/>
+        <import addon="script.module.codequick" version="0.9.5"/>
+        <import addon="script.module.youtube.dl" version="18.0.0"/>
     </requires>
     <extension point="xbmc.python.pluginsource" library="addon.py">
         <provides>video</provides>
@@ -24,6 +25,7 @@
             <screenshot>resources/screenshots/screenshot001.jpg</screenshot>
             <screenshot>resources/screenshots/screenshot002.jpg</screenshot>
         </assets>
-        <news>[fix] Fixed playback problem with Kodi on Xbox.</news>
+        <news>[upd] Add support for reuse of language invoker. [upd] Some small tweaks.</news>
+        <reuselanguageinvoker>true</reuselanguageinvoker>
     </extension>
 </addon>
